<SelectNext /> - first cut
